Updating my repositories README files
We were given this assignment in the Momentum Learning boot camp. At the time I was assigned this project, both knowledge and skill were very limited given that it was close to the beginning of the course. I plan to go back through the project and hone what skills, tools, and knowledge I have learned after the given assignment into a newer, optimal version.